Update letsencrypt_service
Prevent overwriting the domain-symlinks in update_cert{} when the $base_domain is equal to the DEFAULT_HOST environment variable of the nginx-reverseproxy container. This allows using the acme-companion with a DEFAULT_HOST which serves the same host name as other VHOSTS. Prior to this change setting up a host name in more than one VHOST creates a random association between the nginx vhost configuration and one of the available certificates.